[Refactor]: dto 변환로직 repository용 mapper 생성후 해당 객체로 이동.
