+++ src/main/java/com/tenten/linkhub/domain/space/repository/common/mapper/RepositoryDtoMapper.java
@@ -1,22 +1,25 @@
-package com.tenten.linkhub.domain.space.repository.common.dto;
+package com.tenten.linkhub.domain.space.repository.common.mapper;
 
 import com.tenten.linkhub.domain.space.model.space.SpaceImage;
+import com.tenten.linkhub.domain.space.repository.common.dto.SpaceAndOwnerNickName;
+import com.tenten.linkhub.domain.space.repository.common.dto.SpaceAndSpaceImageOwnerNickName;
+import org.springframework.stereotype.Component;
 
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-public record SpaceAndSpaceImageOwnerNickNames(
-        List<SpaceAndSpaceImageOwnerNickName> contents
-) {
-    public static SpaceAndSpaceImageOwnerNickNames of(List<SpaceAndOwnerNickName> spaceAndOwnerNickNames, List<SpaceImage> spaceImages) {
+@Component
+public class RepositoryDtoMapper {
+
+    public List<SpaceAndSpaceImageOwnerNickName> toSpaceAndSpaceImageOwnerNickNames(List<SpaceAndOwnerNickName> spaceAndOwnerNickNames, List<SpaceImage> spaceImages) {
         Map<Long, List<SpaceImage>> spaceImageMap = spaceImages
                 .stream()
                 .collect(Collectors.groupingBy(
                         si -> si.getSpace().getId()
                 ));
 
-        List<SpaceAndSpaceImageOwnerNickName> mapSpaceAndSpaceImageOwnerNickNames = spaceAndOwnerNickNames
+        return spaceAndOwnerNickNames
                 .stream()
                 .map(so -> new SpaceAndSpaceImageOwnerNickName(
                         so.space(),
@@ -24,7 +27,6 @@ public record SpaceAndSpaceImageOwnerNickNames(
                         so.ownerNickName()
                 ))
                 .collect(Collectors.toList());
-
-        return new SpaceAndSpaceImageOwnerNickNames(mapSpaceAndSpaceImageOwnerNickNames);
     }
+
 }
