+++ src/main/resources/application.yml
@@ -16,7 +16,7 @@ spring:
       url: ${LOKI_URL}
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
-    url: ${RDS_URL}
+    url: jdbc:mysql://localhost:3306/link_hub_query_test?serverTimezone=Asia/Seoul&rewriteBatchedStatements=true
     username: ${RDS_USERNAME}
     password: ${RDS_PASSWORD}
   servlet:
